\name{Undocumented functions}
\alias{all_combn}
\alias{allsolution}
\alias{create_sample_path}
\alias{createSamplePath}
\alias{fixed_combn}
\alias{good_path2}
\alias{match_matrix}
\alias{multicore}
\alias{optimize_ga}
\alias{pathBRIDES}
\alias{save_network_big}
\alias{split_sample}
\alias{stepwise}
\title{Undocumented functions}
\description{
  The following functions are for internal computation only: 
     \emph{'all_combn' 'allsolution' 'create_sample_path','createSamplePath','fixed_combn'
     'good_path2' 'match_matrix' 'multicore' 'optimize_ga' 'pathBRIDES'
     'save_network_big' 'split_sample' 'stepwise'} and should not be used independantly.
	 Most of them are helper functions in the calculation of the BRIDES statistics.
}